week3/project/currency_convertor/requirements.txt
requests
python-dotenv
pytest
pytest-cov
